refactor(Navigation): naming Home > HomeNavigation, moving isAuthorised selector to App (as on Navigation seems to throw an error)
