+++ src/components/HomeHeader/HomeHeader.tsx
@@ -1,7 +1,7 @@
 import { StackNavigationProp } from '@react-navigation/stack'
 import { ZIcon } from 'assets/images'
-import { HomeNavigationRoutes } from 'modules/Home/Home.routes'
-import { HomeNavigatorParamsList } from 'modules/Home/Home.types'
+import { HomeNavigationRoutes } from 'modules/HomeNavigation/HomeNavigation.routes'
+import { HomeNavigatorParamsList } from 'modules/HomeNavigation/HomeNavigation.types'
 import React from 'react'
 import { TouchableOpacity, View } from 'react-native'
 import { Colors } from 'styles'
