+++ src/modules/App/App.container.ts
@@ -1,0 +1,9 @@
+import { connect } from 'react-redux'
+
+import App from './App'
+import selector from './App.selector'
+
+const mapStateToProps = selector
+const mapDispatchToProps = null
+
+export default connect(mapStateToProps, mapDispatchToProps)(App)
