+++ src/modules/AuthNavigation/index.ts
@@ -1,0 +1,3 @@
+import AuthNavigation from './AuthNavigation'
+
+export default AuthNavigation
