+++ src/modules/DigitalCv/DigitalCv.tsx
@@ -1,7 +1,7 @@
 import { StackNavigationProp } from '@react-navigation/stack'
 import { ViewContainer, HomeHeader, FirstTimeCard, CvCard } from 'components'
-import { HomeNavigationRoutes } from 'modules/Home/Home.routes'
-import { HomeNavigatorParamsList } from 'modules/Home/Home.types'
+import { HomeNavigationRoutes } from 'modules/HomeNavigation/HomeNavigation.routes'
+import { HomeNavigatorParamsList } from 'modules/HomeNavigation/HomeNavigation.types'
 import React from 'react'
 import { useTranslation } from 'react-i18next'
 import { ScrollView } from 'react-native'
