+++ src/modules/Education/Education.tsx
@@ -1,8 +1,8 @@
 import { StackNavigationProp } from '@react-navigation/stack'
 import { EmptyCard, Card, InfoCard, NormalHeader, Optional, ViewContainer } from 'components'
 import { FormikProps, FormikValues } from 'formik'
-import { HomeNavigationRoutes } from 'modules/Home/Home.routes'
-import { HomeNavigatorParamsList } from 'modules/Home/Home.types'
+import { HomeNavigationRoutes } from 'modules/HomeNavigation/HomeNavigation.routes'
+import { HomeNavigatorParamsList } from 'modules/HomeNavigation/HomeNavigation.types'
 import React, { useRef, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { FlatList, ScrollView } from 'react-native'
