+++ src/modules/HomeNavigation/index.ts
@@ -1,0 +1,3 @@
+import HomeNavigation from './HomeNavigation'
+
+export default HomeNavigation
