--- src/modules/Navigation/Navigation.container.tsx
@@ -1,9 +1,0 @@
-import { connect } from 'react-redux'
-
-import Navigation from './Navigation'
-import selector from './Navigation.selector'
-
-const mapStateToProps = selector
-const mapDispatchToProps = null
-
-export default connect(mapStateToProps, mapDispatchToProps)(Navigation)
