+++ src/modules/Navigation/Navigation.tsx
@@ -3,7 +3,7 @@ import { createStackNavigator } from '@react-navigation/stack'
 import React from 'react'
 
 import AuthNavigation from '../AuthNavigation/AuthNavigation'
-import Home from '../Home/Home'
+import HomeNavigation from '../HomeNavigation'
 import linking from './Linking'
 
 const Stack = createStackNavigator()
@@ -16,7 +16,7 @@ const Navigation = ({ isAuthorised }: Props) => (
   <NavigationContainer linking={linking}>
     <Stack.Navigator headerMode="none">
       {isAuthorised ? (
-        <Stack.Screen name="Home" component={Home} />
+        <Stack.Screen name="Home" component={HomeNavigation} />
       ) : (
         <Stack.Screen name="Authentication" component={AuthNavigation} />
       )}
