+++ src/modules/Profile/Profile.form.tsx
@@ -2,8 +2,8 @@ import { StackNavigationProp } from '@react-navigation/stack'
 import { DropDown, Input, Spinner } from 'components'
 import countries from 'constants/countries'
 import { Formik, FormikProps, FormikValues } from 'formik'
-import { HomeNavigationRoutes } from 'modules/Home/Home.routes'
-import { HomeNavigatorParamsList } from 'modules/Home/Home.types'
+import { HomeNavigationRoutes } from 'modules/HomeNavigation/HomeNavigation.routes'
+import { HomeNavigatorParamsList } from 'modules/HomeNavigation/HomeNavigation.types'
 import React, { forwardRef, useImperativeHandle, useRef } from 'react'
 import { useTranslation } from 'react-i18next'
 import { mapToDropDownArray } from 'utils/strings.utils'
