+++ src/modules/Skills/Skills.tsx
@@ -1,8 +1,8 @@
 import { StackNavigationProp } from '@react-navigation/stack'
 import { Card, EmptyCard, NormalHeader, Optional, SkillCard, ViewContainer } from 'components'
 import ListCardHeader from 'components/ListCardHeader'
-import { HomeNavigationRoutes } from 'modules/Home/Home.routes'
-import { HomeNavigatorParamsList } from 'modules/Home/Home.types'
+import { HomeNavigationRoutes } from 'modules/HomeNavigation/HomeNavigation.routes'
+import { HomeNavigatorParamsList } from 'modules/HomeNavigation/HomeNavigation.types'
 import React, { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { FlatList } from 'react-native'
